added update email(bug) and password

+++ src/main/java/com/esprit/cloudcraft/repository/TokenRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token,Integer> {
+public interface TokenRepository extends JpaRepository<Token,Long> {
 
     @Query("""
         select t from Token t inner join User u on t.user.id = u.id 
         where u.id = :userId  and (t.expired = false or t.revoked = false)
 """)
-    List<Token> findAllValidTokensByUser(Integer userId);
+    List<Token> findAllValidTokensByUser(Long userId);
 
     Optional<Token> findByToken(String token);
 }
